Squashed commits on 373
Specialized use of Tkinter and stopped passing wx as argument:

The basic idea is to first load bolt for the localization of error
messages and then load wxpython. If any of the two raise an exception
then we use Tkinter to show the error message. If those two are
successfully imported then we know wx is available and we do not need
to pass it between different functions.

Change bash debug.bat to not crash when opened a second time:

Wrye Bash Debug.bat would crash or more correctly close without opening
a window or printing an error message if there was another instance of
Wrye Bash started by Wrye Bash Debug.bat already running.

Merge _showErrorInAnyGui and _showErrorInGui into _show_wx_error:

The refactored usage of TKinter allowed for merging the two error-
message functions into a single one that offers the full fuctionality.