Specialized use of Tkinter and stopped passing wx as argument:
The basic idea is to first load bolt for the localization of error
messages and then load wxpython. If any of the two raise an exception
then we use Tkinter to show the error message. If those two are
successfully imported then we know wx is available and we do not need
to pass it between different functions.

Changes the line endings of Bash Readme Template.html because they are
CRLF on the github repo, which should not happen. This creates a issue
where git constantly sees unstaged changes, where no changes where made
and would not let me finish the rebase without also changing this file.